+++ src/core/shared/async-function.ts
@@ -1,0 +1,1 @@
+export const AsyncFunction = (async () => {}).constructor;