+++ src/core/shared/domain-event.ts
@@ -1,0 +1,11 @@
+export abstract class DomainEvent {
+  private occuredOn: Date;
+
+  constructor(public readonly name: string) {
+    this.occuredOn = new Date();
+  }
+
+  public getOccuredOn() {
+    return this.occuredOn;
+  }
+}