+++ src/core/shared/domain-subscriber.ts
@@ -1,0 +1,5 @@
+import { DomainEvent } from './domain-event';
+
+export abstract class DomainSubscriber<T extends DomainEvent> {
+  public abstract setupSubscriptions(): void;
+}